split fritzhomedevice into smaller classes to improve maintainability
