+++ pyfritzhome/__init__.py
@@ -1,12 +1,10 @@
 from .errors import InvalidError, LoginError
 from .fritzhome import Fritzhome
 from .fritzhomedevice import FritzhomeDevice
-from .fritzhomedevicefeatures import FritzhomeDeviceFeatures
 
 __all__ = [
     "Fritzhome",
     "FritzhomeDevice",
-    "FritzhomeDeviceFeatures",
     "InvalidError",
     "LoginError",
 ]
