+++ pyfritzhome/devicetypes/__init__.py
@@ -1,0 +1,17 @@
+from .fritzhomedevicealarm import FritzhomeDeviceAlarm
+from .fritzhomedevicebutton import FritzhomeDeviceButton
+from .fritzhomedevicepowermeter import FritzhomeDevicePowermeter
+from .fritzhomedevicerepeater import FritzhomeDeviceRepeater
+from .fritzhomedeviceswitch import FritzhomeDeviceSwitch
+from .fritzhomedevicetemperature import FritzhomeDeviceTemperature
+from .fritzhomedevicethermostat import FritzhomeDeviceThermostat
+
+__all__ = [
+    "FritzhomeDeviceAlarm",
+    "FritzhomeDeviceButton",
+    "FritzhomeDevicePowermeter",
+    "FritzhomeDeviceRepeater",
+    "FritzhomeDeviceSwitch",
+    "FritzhomeDeviceTemperature",
+    "FritzhomeDeviceThermostat",
+]
